breaks user and role folders out of users and folders into separate listings
