+++ config/dao-classmap.php
@@ -145,11 +145,13 @@ return [
     'Pimcore\\Model\\User' => 'Pimcore\\Model\\User\\Dao',
     'Pimcore\\Model\\User\\AbstractUser' => 'Pimcore\\Model\\User\\AbstractUser\\Dao',
     'Pimcore\\Model\\User\\Folder' => 'Pimcore\\Model\\User\\Dao',
+    'Pimcore\\Model\\User\\Folder\\Listing' => 'Pimcore\\Model\\User\\Folder\\Listing\\Dao',
     'Pimcore\\Model\\User\\Listing' => 'Pimcore\\Model\\User\\Listing\\Dao',
     'Pimcore\\Model\\User\\Permission\\Definition' => 'Pimcore\\Model\\User\\Permission\\Definition\\Dao',
     'Pimcore\\Model\\User\\Permission\\Definition\\Listing' => 'Pimcore\\Model\\User\\Permission\\Definition\\Listing\\Dao',
     'Pimcore\\Model\\User\\Role' => 'Pimcore\\Model\\User\\Role\\Dao',
     'Pimcore\\Model\\User\\Role\\Folder' => 'Pimcore\\Model\\User\\Role\\Folder\\Dao',
+    'Pimcore\\Model\\User\\Role\\Folder\\Listing' => 'Pimcore\\Model\\User\\Role\\Folder\\Listing\\Dao',
     'Pimcore\\Model\\User\\Role\\Listing' => 'Pimcore\\Model\\User\\Role\\Listing\\Dao',
     'Pimcore\\Model\\User\\UserRole' => 'Pimcore\\Model\\User\\UserRole\\Dao',
     'Pimcore\\Model\\User\\UserRole\\Folder' => 'Pimcore\\Model\\User\\UserRole\\Dao',
