+++ models/User/Role/Folder/Listing/Dao.php
@@ -1,0 +1,25 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * This source file is available under the terms of the
+ * Pimcore Open Core License (POCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (https://www.pimcore.com)
+ *  @license    Pimcore Open Core License (POCL)
+ */
+
+namespace Pimcore\Model\User\Role\Folder\Listing;
+
+use Pimcore\Model;
+
+/**
+ * @internal
+ *
+ * @property \Pimcore\Model\User\Role\Folder\Listing $model
+ */
+class Dao extends Model\User\Listing\AbstractListing\Dao
+{
+}
